stores patient data into an array

+++ get_data.sh
@@ -1,43 +1,30 @@
 #!/bin/bash
 
-function getPatient {
-	patient_id=$1
-	url=$HOST:$PORT/rest/families/$patient_id
-
-	local patient = $(curlToPheno url="$HOST:$PORT/rest/patients/$patient_id")
-	echo $patient
-}
-
-function getFamily {
-	family_id=$1
-	url=$HOST:$PORT/rest/families/$family_id
-
-	local family = $(curlToPheno $url)
-	echo $family
-}
-
 function curlToPheno {
-	url=$1
+	url=$HOST:$PORT/rest/$1/$2
 	args="-u $USER:$PSWD -X GET $url"
 
+	request="curl -s $args"
 	status=$(curl -o .temp -s -w "%{http_code}\n" $args)
 	
-	echo $status
-	if [[ $status -eq 200 ]]; then
+	if [[ $status -eq 200 ]]; then	
 		response=$($request)
-		echo $response
 	else
 		response=$status
-		echo $status
 	fi;
 
-	rm .temp
+	if [ -e .temp ]; then 
+		rm .temp
+	fi;
+	
+	echo $response
 	loggit "$request" "$response"
 }
 
 function loggit {
-	echo "Logging curl results to" $LOGFILE
     echo $(date '+%d/%m/%Y %H:%M:%S') >> $LOGFILE
     echo "Command:" $1 >> $LOGFILE
     echo $2 $'\r' $'\r' >> $LOGFILE
 }
+
+
